+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 All notable changes to the "sshextension" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
- 
+
+## 0.0.5 (Features)
+- Added name for instance of integrated terminal
+- Added check for the existence of the ssh utility on your machine
+- For an already open server, a new instance of the integrated terminal is not created
+- Fixed loading of the ftp-simple configuration file when it was changed
+
 ## 0.0.4 (Patch)
 - Fixed bug with no ftp-simple.json file